BuyList component unit tests done
